fix a very stupid way of addressing memory which is not 64bit compatible in the miptex reader
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@4587 d7cf8633-e32d-0410-b094-e92efae38249

+++ wad.c
@@ -209,7 +209,7 @@ qbyte *W_ConvertWAD3Texture(miptex_t *tex)
 	qbyte *in, *data, *out, *pal;
 	int d, p;
 
-	in = (qbyte *)((int) tex + tex->offsets[0]);
+	in = (qbyte *)tex + tex->offsets[0];
 	data = out = Mem_Alloc(tempmempool, tex->width * tex->height * 4);
 	if (!data)
 		return NULL;
